feat: statically generate projects page
