+++ go.mod
@@ -2,4 +2,7 @@ module github.com/sioodmy/website
 
 go 1.22.2
 
-require github.com/gomarkdown/markdown v0.0.0-20240419095408-642f0ee99ae2 // indirect
+require (
+	github.com/gomarkdown/markdown v0.0.0-20240419095408-642f0ee99ae2 // indirect
+	github.com/pelletier/go-toml v1.9.5 // indirect
+)
