+++ templates/post.html
@@ -21,11 +21,11 @@
     {{.Html}}
   </main>
   <nav>
-    <a href="index.html">About</a>
-    <a href="projects.html">Projects</a>
-    <a href="contact.html">Contact</a>
-    <a href="donate.html">Donate</a>
-    <a href="blog.html" class="selected">Blog</a>
+    <a href="../index.html">About</a>
+    <a href="../projects.html">Projects</a>
+    <a href="../contact.html">Contact</a>
+    <a href="../donate.html">Donate</a>
+    <a href="../blog.html" class="selected">Blog</a>
   </nav>
 </body>
