+++ templates/projects.html
@@ -18,27 +18,14 @@
 
 <body>
   <div class="container projects">
+    {{range .Projects}}
     <div class="project">
-      <a href="https://github.com/schizofox/schizofox/" target="_blank">
-        <h3>Schizofox</h3>
+      <a href="{{.RepoUrl}}" target="_blank">
+        <h3>{{.Name}}</h3>
       </a>
-      Hardened Firefox nix flake with our custom declarative configuration features like extension settings and theming
-      focused on privacy and security while also maintaining usability and compatibility with modern web unlike other
-      similar projects
+      {{.Description}}
     </div>
-    <div class="project">
-      <a href="https://github.com/sioodmy/schizosearch" target="_blank">
-        <h3>Schizosearch</h3>
-      </a>
-      Minimal privacy respecting meta search engine written in Go (previously in rust)
-    </div>
-    <div class="project">
-      <a href="https://github.com/sioodmy/dotfiles" target="_blank">
-        <h3>Dotfiles</h3>
-      </a>
-      Starting point of my Nix experience. NixOS configuration flake for all of my devices and server infrastracture
-      with built in impermanence, encryption and more
-    </div>
+    {{end}}
   </div>
 
   <nav>
